refactor omegalul
